Make `absolute_directory` method on `ManifestPath`
